Replaces markdown component w/ existing component Moves markdown web component into shared Moves markdown-related functions to a dedicated file
