+++ src/git/remotes/github.ts
@@ -9,7 +9,8 @@ import type { Brand, Unbrand } from '../../system/brand';
 import { fromNow } from '../../system/date';
 import { memoize } from '../../system/decorators/memoize';
 import { encodeUrl } from '../../system/encoding';
-import { equalsIgnoreCase, escapeMarkdown, unescapeMarkdown } from '../../system/string';
+import { escapeMarkdown, unescapeMarkdown } from '../../system/markdown';
+import { equalsIgnoreCase } from '../../system/string';
 import { getIssueOrPullRequestMarkdownIcon } from '../models/issue';
 import { isSha } from '../models/reference';
 import type { Repository } from '../models/repository';
