+++ src/system/markdown.ts
@@ -1,10 +1,39 @@
+const escapeMarkdownRegex = /[\\*_{}[\]()#+\-.!]/g;
+const unescapeMarkdownRegex = /\\([\\`*_{}[\]()#+\-.!])/g;
+
+const escapeMarkdownHeaderRegex = /^===/gm;
+const unescapeMarkdownHeaderRegex = /^\u200b===/gm;
+
+// const sampleMarkdown = '## message `not code` *not important* _no underline_ \n> don\'t quote me \n- don\'t list me \n+ don\'t list me \n1. don\'t list me \nnot h1 \n=== \nnot h2 \n---\n***\n---\n___';
+const markdownQuotedRegex = /\r?\n/g;
+const markdownBacktickRegex = /`/g;
+
+export function escapeMarkdown(s: string, options: { quoted?: boolean; inlineBackticks?: boolean } = {}): string {
+	s = s
+		// Escape markdown
+		.replace(escapeMarkdownRegex, '\\$&')
+		// Escape markdown header (since the above regex won't match it)
+		.replace(escapeMarkdownHeaderRegex, '\u200b===');
+
+	if (options.inlineBackticks) {
+		s = escapeMarkdownCodeBlocks(s);
+	} else {
+		s = s.replace(markdownBacktickRegex, '\\$&');
+	}
+	if (!options.quoted) return s;
+
+	// Keep under the same block-quote but with line breaks
+	return s.trim().replace(markdownQuotedRegex, '\t\\\n>  ');
+}
+
 /**
  * escapes markdown code blocks
  */
-function escapeTripleBackticks(s: string) {
+export function escapeMarkdownCodeBlocks(s: string) {
 	const tripleBackticks = '```';
 	const escapedTripleBackticks = '\\`\\`\\`';
-	let str = '';
+
+	let result = '';
 	let allowed = true;
 	let quotesOpened = false;
 	let buffer = '';
@@ -19,9 +48,9 @@ function escapeTripleBackticks(s: string) {
 			if (quotesOpened) {
 				quotesOpened = false;
 				if (allowed) {
-					str += `${tripleBackticks}${buffer}${tripleBackticks}`;
+					result += `${tripleBackticks}${buffer}${tripleBackticks}`;
 				} else {
-					str += `${escapedTripleBackticks}${buffer}${escapedTripleBackticks}`;
+					result += `${escapedTripleBackticks}${buffer}${escapedTripleBackticks}`;
 					allowed = true;
 				}
 				buffer = '';
@@ -35,10 +64,24 @@ function escapeTripleBackticks(s: string) {
 		if (quotesOpened) {
 			buffer += char;
 		} else {
-			str += char;
+			result += char;
 		}
 	}
-	return str;
+
+	if (quotesOpened) {
+		// Handle unclosed code block
+		result += allowed ? tripleBackticks + buffer : escapedTripleBackticks + buffer;
+	}
+
+	return result;
 }
 
-export { escapeTripleBackticks };
+export function unescapeMarkdown(s: string): string {
+	return (
+		s
+			// Unescape markdown
+			.replace(unescapeMarkdownRegex, '$1')
+			// Unescape markdown header
+			.replace(unescapeMarkdownHeaderRegex, '===')
+	);
+}
