+++ src/webviews/apps/plus/graph/GraphWrapper.tsx
@@ -71,6 +71,7 @@ import { createCommandLink } from '../../shared/commands';
 import { GlButton } from '../../shared/components/button.react';
 import { CodeIcon } from '../../shared/components/code-icon.react';
 import { GlConnect } from '../../shared/components/integrations/connect.react';
+import { GlMarkdown } from '../../shared/components/markdown/markdown.react';
 import { MenuDivider, MenuItem, MenuLabel, MenuList } from '../../shared/components/menu/react';
 import { PopMenu } from '../../shared/components/overlays/pop-menu/react';
 import { GlPopover } from '../../shared/components/overlays/popover.react';
@@ -83,7 +84,6 @@ import type { SearchNavigationEventDetail } from '../../shared/components/search
 import type { DateTimeFormat } from '../../shared/date';
 import { formatDate, fromNow } from '../../shared/date';
 import { GlGraphHover } from './hover/graphHover.react';
-import { Markdown } from './markdown/markdown';
 import type { GraphMinimapDaySelectedEventDetail } from './minimap/minimap';
 import { GlGraphMinimapContainer } from './minimap/minimap-container.react';
 import { GlGraphSideBar } from './sidebar/sidebar.react';
@@ -1606,7 +1606,7 @@ export function GraphWrapper({
 							columnsSettings={columns}
 							contexts={context}
 							// @ts-expect-error returnType of formatCommitMessage callback expects to be string, but it works fine with react element
-							formatCommitMessage={e => <Markdown>{e}</Markdown>}
+							formatCommitMessage={e => <GlMarkdown markdown={e}></GlMarkdown>}
 							cssVariables={styleProps?.cssVariables}
 							dimMergeCommits={graphConfig?.dimMergeCommits}
 							downstreamsByUpstream={downstreams}
